Update for Sorting_Algorithms.py
Created 3 functions that generates lists of numbers from 1 to k. In order, reverse ordered, and random order.
Created new performance test functions to test bubble, insert, and merge sorts for the best, average, and worst case times. 
Created averaging functions that run the performance test 100 times in a loop and gets the average of those times.